requirements.txt
pyyaml
psycopg2-binary
sqlalchemy  # required by APScheduler SQLAlchemyJobStore (scheduler.py only)
apscheduler==3.10.4
yfinance==0.2.66
curl-cffi==0.7.4
google-generativeai==0.8.3
anthropic
python-dotenv==1.0.0
requests==2.31.0
lxml==5.3.0